+++ odtools/odlivefeed/src/LiveFeed.cpp
@@ -20,17 +20,10 @@
 #include <iostream>
 #include <sstream>
 
-#ifdef HAVE_DL
-    #include <dlfcn.h>
-    #include <experimental/filesystem>
-#endif
-
 #include <opendavinci/odcore/data/Container.h>
 #include <opendavinci/odcore/data/TimeStamp.h>
 #include <opendavinci/odcore/reflection/Message.h>
 #include <opendavinci/odcore/strings/StringToolbox.h>
-#include <opendavinci/GeneratedHeaders_OpenDaVINCI_Helper.h>
-#include <automotivedata/GeneratedHeaders_AutomotiveData_Helper.h>
 
 #include "LiveFeed.h"
 
@@ -41,149 +34,13 @@ namespace odlivefeed {
     using namespace odcore::data;
     using namespace odcore::reflection;
 
-    ////////////////////////////////////////////////////////////////////////////
-
-    HelperEntry::HelperEntry() : 
-         m_library(""),
-         m_dynamicObjectHandle(NULL),
-         m_helper(NULL)
-    {}
-
-    HelperEntry::HelperEntry(const HelperEntry &obj) :
-        m_library(obj.m_library),
-        m_dynamicObjectHandle(obj.m_dynamicObjectHandle),
-        m_helper(obj.m_helper)
-    {}
-
-    HelperEntry& HelperEntry::operator=(const HelperEntry &obj) {
-        m_library = obj.m_library;
-        m_dynamicObjectHandle = obj.m_dynamicObjectHandle;
-        m_helper = obj.m_helper;
-        return *this;
-    }
-
-    HelperEntry::~HelperEntry() {}
-
-    ////////////////////////////////////////////////////////////////////////////
-
-
     LiveFeed::LiveFeed(const int32_t &argc, char **argv) :
         DataTriggeredConferenceClientModule(argc, argv, "odlivefeed"),
         m_mainwindow(NULL),
         m_mapOfLastContainers(),
-        m_listOfLibrariesToLoad(),
-        m_listOfHelpers() {
-        const string SEARCH_PATH = "/opt";
-        const vector<string> paths = odcore::strings::StringToolbox::split(SEARCH_PATH, ',');
-        m_listOfLibrariesToLoad = getListOfLibrariesToLoad(paths);
+        m_messageResolver({"/opt"}, "libodvd", ".so") {}
 
-        findAndLoadSharedLibraries();
-    }
-
-    LiveFeed::~LiveFeed() {
-        unloadSharedLibraries();
-    }
-
-
-    vector<string> LiveFeed::getListOfLibrariesToLoad(const vector<string> &paths) {
-        vector<string> librariesToLoad;
-#ifndef HAVE_DL
-        (void)paths;
-#endif
-
-#ifdef HAVE_DL
-        for(auto pathToSearch : paths) {
-            try {
-                for(auto &pathElement : std::experimental::filesystem::recursive_directory_iterator(pathToSearch)) {
-                    stringstream sstr;
-                    sstr << pathElement;
-                    string entry = sstr.str();
-                    if (entry.find("libodvd") != string::npos) {
-                        if (entry.find(".so") != string::npos) {
-                            vector<string> path = odcore::strings::StringToolbox::split(entry, '/');
-                            if (path.size() > 0) {
-                                string lib = path[path.size()-1];
-                                if (lib.size() > 0) {
-                                    lib = lib.substr(0, lib.size()-1);
-                                    librariesToLoad.push_back(lib);
-                                }
-                            }
-                        }
-                    }
-                }
-            }
-            catch(...) {}
-        }
-#endif
-        return librariesToLoad;
-    }
-
-    void LiveFeed::findAndLoadSharedLibraries() {
-#ifdef HAVE_DL
-        auto it = m_listOfLibrariesToLoad.begin();
-        while (it != m_listOfLibrariesToLoad.end()) {
-            const string libraryToLoad = *it;
-
-            {
-                HelperEntry e;
-
-                cout << "[LiveFeed] Opening '" + libraryToLoad + "'..." << endl;
-                e.m_dynamicObjectHandle = dlopen(libraryToLoad.c_str(), RTLD_LAZY);
-
-                if (!e.m_dynamicObjectHandle) {
-                    cerr << "[LiveFeed] Cannot open library '" + libraryToLoad + "': " << dlerror() << endl;
-                }
-                else {
-                    typedef odcore::reflection::Helper *createHelper_t();
-
-                    // reset errors
-                    dlerror();
-                    createHelper_t* getHelper = (createHelper_t*) dlsym(e.m_dynamicObjectHandle, "newHelper");
-                    const char *dlsym_error = dlerror();
-                    if (dlsym_error) {
-                        cerr << "[LiveFeed] Cannot load symbol 'newHelper' from '" + libraryToLoad + "': " << dlsym_error << endl;
-                        dlclose(e.m_dynamicObjectHandle);
-                    }
-                    else {
-                        // Get pointer to external message handling.
-                        e.m_helper = getHelper();
-                        e.m_library = libraryToLoad;
-                        m_listOfHelpers.push_back(e);
-                    }
-                }
-            }
-
-            it++;
-        }
-#endif
-    }
-
-    void LiveFeed::unloadSharedLibraries() {
-#ifdef HAVE_DL
-        auto it = m_listOfHelpers.begin();
-        while (it != m_listOfHelpers.end()) {
-            HelperEntry e = *it;
-
-            // Type to refer to the destroy method inside the shared library.
-            typedef void deleteHelper_t(odcore::reflection::Helper *);
-
-            // Reset error messages from dynamically loading shared object.
-            dlerror();
-            deleteHelper_t* delHelper = (deleteHelper_t*) dlsym(e.m_dynamicObjectHandle, "deleteHelper");
-            const char *dlsym_error = dlerror();
-            if (dlsym_error) {
-                cerr << "[LiveFeed] Cannot load symbol 'deleteHelper': " << dlsym_error << endl;
-            }
-            else {
-                cout << "[LiveFeed] Closing link to '" + e.m_library + "'" << endl;
-                delHelper(e.m_helper);
-            }
-            dlclose(e.m_dynamicObjectHandle);
-
-            it++;
-        }
-#endif
-    }
+    LiveFeed::~LiveFeed() {}
 
     void LiveFeed::setUp() {
         m_mainwindow = initscr();
@@ -227,21 +84,7 @@ namespace odlivefeed {
 
                     // Try to extract messages.
                     bool successfullyMapped = false;
-                    // First, try to decode a regular OpenDaVINCI message.
-                    odcore::reflection::Message msg = GeneratedHeaders_OpenDaVINCI_Helper::__map(containerEntry, successfullyMapped);
-
-                    // Next, try to decode a regular OpenDaVINCI message.
-                    if (!successfullyMapped) {
-                        msg = GeneratedHeaders_AutomotiveData_Helper::__map(containerEntry, successfullyMapped);
-                    }
-
-                    // Try dynamically loaded libraries next.
-                    auto kt = m_listOfHelpers.begin();
-                    while ( (!successfullyMapped) && (kt != m_listOfHelpers.end())) {
-                        HelperEntry e = *kt;
-                        msg = e.m_helper->map(containerEntry, successfullyMapped);
-                        kt++;
-                    }
+                    odcore::reflection::Message msg = m_messageResolver.resolve(containerEntry, successfullyMapped);
 
                     if (successfullyMapped) {
                         sstr << "Container: " << containerEntry.getDataType() << "/" << containerEntry.getSenderStamp() << ": " << " Sent: " << containerEntry.getSentTimeStamp().getYYYYMMDD_HHMMSSms() << ", sample time: " << containerEntry.getSampleTimeStamp().getYYYYMMDD_HHMMSSms() << ": " << msg.getLongName();
@@ -265,3 +108,4 @@ namespace odlivefeed {
     }
 
 } // odlivefeed
+
